restructure middleware
