--- lib/sitehub/logging/log_stash.rb
@@ -1,10 +1,0 @@
-require_relative 'log_entry'
-class SiteHub
-  module Logging
-    class LogStash < Array
-      def <<(message)
-        super(LogEntry.new(message))
-      end
-    end
-  end
-end
