+++ lib/sitehub/middleware/logging.rb
