--- spec/sitehub/logging/log_stash_spec.rb
@@ -1,19 +1,0 @@
-require 'sitehub/logging/log_stash'
-
-class SiteHub
-  module Logging
-    describe LogStash do
-      it 'inherits Array' do
-        expect(subject).to be_a_kind_of(Array)
-      end
-
-      describe '#<<' do
-        it 'adds a LogEntry' do
-          allow(Time).to receive(:now).and_return(:current_time)
-          subject << :message
-          expect(subject).to eq([LogEntry.new(:message)])
-        end
-      end
-    end
-  end
-end
